Fix privacy provider signature (fatal error)
The privacy provider get_reason function must have return type : string or it will cause a fatal error (when run on PHP 7). This breaks the entire privacy API on the system so it's fairly important. :)

The simplest fix is to add ':string' in the function definition, but if you want it to also work on Moodle 3.3 (the only version that supports the privacy API, but allows a PHP version less than 7) then the documentation says you need the more complicated fix using legacy polyfill, which I've done here.