Login in dockerhub when possible to avoid pull limits
With current pull limits @ dockerhub, it's possible to get failed
jobs just running 2-3 builds (each one having a good number of jobs).

With this patch, if both DOCKER_USER and DOCKER_TOKEN are defined
as environmental variables, they will be used to perform a login
to dockerhub, getting rid of pull limits.