Use exttests both for http and https
It was promised @ #200 to fix #178, so here it is the fix.

Basically, without needing to put the exttests container with
a proper https configuration at all, use the existing http one
for the tests requiring https (because it's not checked).

We use the very same trick @ CIs and they are working and
running all the tests without a problem.

Fixes #178