Bump MySQL default to 8.0 and PostgreSQL to 13
They are the lowest versions supported for 4.2 and up.

Note that, at the time of writing this that's just a proposal, but that
way we can test if they work for all supported branches in advance.

So, here, we make MySQL 8.0 the default and create specific versions
for 5.6 and 5.7 in case anybody wants to use them.