Prepare app config for 4.4.0
The 4.4.0 release of the app changes some things that need to be updated in the Docker images. In particular, moving away from WebSQL (MOBILE-4304) introduced a requirement to run the application in a secure context; and the app will run with the https:// protocol on port 443 by default. The app also needs to run in a newer version of Chrome now; so when the app is being used it will default to version 120.0.

These changes will be applied for the ionic7 runtime, which also changed in this release. This update to moodle-docker removes all the code related with the ionic3 runtime, which hasn't been used in the app since version 3.9.4 (December 2020). Given that this should work out of the box now, the documentation for the MOODLE_DOCKER_APP_RUNTIME and MOODLE_DOCKER_APP_NODE_VERSION variables has been removed.