Reconfigure the PHP Handler to be conditional on file existence
Set the handler to use for files whose extension ends in `.php`.
This is a workaround for https://github.com/docker-library/php/issues/1576
The `FallbackResource` is a default Handler action which is only used if a Handler has not already been applied.
The docker/php image sets the handler unconditionally for any file whose name matches `.php`.
This configuration block unsets the default handler, and instead only applies it if the file exists.